Even more bytecode optimization

+++ battlecode23-scaffold-main/src/qp1/common/SpreadSettings.java
@@ -5,11 +5,11 @@ public enum SpreadSettings {
     LAUNCHER(16, 1, 20);
 
     public final int ally_dist_cutoff;
-    public final int ally_dist_factor;
+    public final float ally_dist_factor;
     public final int random_cutoff;
     public final int random_bound;
 
-    SpreadSettings(int ally_dist_cutoff, int ally_dist_factor, int random_cutoff) {
+    SpreadSettings(int ally_dist_cutoff, float ally_dist_factor, int random_cutoff) {
         this.ally_dist_cutoff = ally_dist_cutoff;
         this.ally_dist_factor = ally_dist_factor;
         this.random_cutoff = random_cutoff;
